Separate test_cli.py
